# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rest Api documentation",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://rest-api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/businesses": {
            "get": {
                "tags": [
                    "Businesses"
                ],
                "summary": "Get list of businesses",
                "description": "Returns list of businesses",
                "operationId": "getBusinessesList",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search businesses by query string value.",
                        "required": false
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort businesses by query string value.",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "enum",
                                "enum": [
                                    "business.name asc",
                                    "business.name desc",
                                    "description asc",
                                    "description desc",
                                    "raiting asc",
                                    "raiting desc",
                                    "category.name asc",
                                    "category.name desc"
                                ]
                            }
                        },
                        "example": "raiting asc;business.name desc"
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter businesses by query string value. (by category.id, by raiting)",
                        "required": false,
                        "example": "category.id 1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessListResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "Businesses"
                ],
                "summary": "Store new business",
                "description": "Returns business data",
                "operationId": "storeBusiness",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreBusinessRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequestResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/api/businesses/{id}": {
            "get": {
                "tags": [
                    "Businesses"
                ],
                "summary": "Get business information",
                "description": "Returns business data",
                "operationId": "getBusinessById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Business id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequestResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "Businesses"
                ],
                "summary": "Update existing business",
                "description": "Returns updated business data",
                "operationId": "updateBusiness",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Business id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateBusinessRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequestResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Businesses"
                ],
                "summary": "Delete existing business",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteBusiness",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Business id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/api/businesses/statistics": {
            "get": {
                "tags": [
                    "Businesses"
                ],
                "summary": "Get count of businesses per day",
                "description": "Returns count of businesses",
                "operationId": "getCountOfBusinessesPerDay",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Start date of period",
                        "required": true,
                        "example": "2020-08-02"
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "End date of period",
                        "required": true,
                        "example": "2020-08-03"
                    },
                    {
                        "name": "owner_id",
                        "in": "query",
                        "description": "Filtering by owner id (available only for admin)",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessCountPerDayResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BusinessListResource": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/SuccessResource"
                    },
                    {
                        "title": "BusinessListResource",
                        "description": "Business list resource",
                        "properties": {
                            "data": {
                                "title": "Data",
                                "description": "Data wrapper",
                                "properties": {
                                    "list": {
                                        "title": "Business list",
                                        "description": "Business list wrapper",
                                        "$ref": "#/components/schemas/BusinessList"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object",
                        "xml": {
                            "name": "BusinessListResource"
                        }
                    }
                ]
            },
...
            "ForbiddenResource": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ErrorResource"
                    },
                    {
                        "title": "ForbiddenResource",
                        "description": "Forbidden resource",
                        "properties": {
                            "message": {
                                "title": "Message",
                                "description": "Message wrapper",
                                "example": "Forbidden"
                            }
                        },
                        "type": "object",
                        "xml": {
                            "name": "ForbiddenResource"
                        }
                    }
                ]
            },
...
            "UnauthenticatedResource": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ErrorResource"
                    },
                    {
                        "title": "UnauthenticatedResource",
                        "description": "Unauthenticated resource",
                        "properties": {
                            "status": {
                                "title": "Status",
                                "description": "Status wrapper",
                                "example": "failed"
                            },
                            "message": {
                                "title": "Message",
                                "description": "Message wrapper",
                                "example": "Unauthenticated"
                            }
                        },
                        "type": "object",
                        "xml": {
                            "name": "UnauthenticatedResource"
                        }
                    }
                ]
            },
...
            "StoreBusinessRequest": {
                "title": "Store Business request",
                "description": "Store Business request body data",
                "required": [
                    "name",
                    "description",
                    "category_id",
                    "raiting",
                    "phone",
                    "email",
                    "website"
                ],
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Name of the business",
                        "type": "string",
                        "example": "This is business name"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the business",
                        "type": "string",
                        "example": "This is business description"
                    },
                    "categoryId": {
                        "title": "category_id",
                        "description": "Id of the category",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "raiting": {
                        "title": "Owner ID",
                        "description": "Number between 0 and 100",
                        "type": "integer",
                        "format": "int64",
                        "example": 75
                    },
                    "phone": {
                        "title": "Phone",
                        "description": "Phone of the business",
                        "type": "string",
                        "example": "+1 (855) 451-9378"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Contact email",
                        "type": "string",
                        "example": "hackett.hector@mante.biz"
                    },
                    "website": {
                        "title": "Website",
                        "description": "Website of the business",
                        "type": "string",
                        "example": "hackett.hector@mante.biz"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://rest-api/oauth/authorize",
                        "tokenUrl": "http://rest-api/oauth/token",
                        "refreshUrl": "http://rest-api/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}